libx86: introduce helper to fetch cpuid leaf
Introduce a helper based on the current Xen guest_cpuid code in order
to fetch a cpuid leaf from a policy. Use such helper in order to
replace the code in x86_cpuid_copy_from_buffer. Note the introduced
helper should not be directly called and instead x86_cpuid_get_leaf
should be used that will properly deal with const and non-const
inputs.

Also add a test to check that the introduced helper doesn't go over
the bounds of the policy.

Suggested-by: Andrew Cooper <andrew.cooper3@citrix.com>
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Changes since v3:
 - New in this version.