tools/configure: make iPXE dependent on QEMU traditional
iPXE is only used by QEMU traditional, so make it off by default
unless QEMU traditional is enabled.

Reported-by: Andrew Cooper <andrew.cooper3@citrix.com>
Fixes: bcf77ce510 ('configure: modify default of building rombios')
Signed-off-by: Roger Pau Monné <roger.,pau@citrix.com>
---
Cc: Ian Jackson <iwj@xenproject.org>

Please run autoconf before commit.

Without this patch FreeBSD build is broken, as iPXE won't build. The
main risk would be getting the configure script runes wrong, but I
think that's fairly easy to check as it will either work on not. It's
a change in the defaults, but it basically restores the behavior
previous to disabling qemu-trad on Linux.