livepatch: do not ignore sections with 0 size
A sideeffect of ignoring such sections is that symbols belonging to
them won't be resolved, and that could make relocations belonging to
other sections fail.

For example it's likely to have an empty .altinstr_replacement with
symbools pointing to it, and marking the section as ignored will
prevent the symbols from being resolved, which in turn will cause any
relocations against them to fail.

In order to solve this do not ignore sections with 0 size, only ignore
sections that don't have the SHF_ALLOC flag set.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>