tools/firmware: do not add a .note.gnu.property section
Prevent the assembler from creating a .note.gnu.property section on
the output objects, as it's not useful for firmware related binaries,
and breaks the resulting rombios image.

This requires modifying the cc-option Makefile macro so it can test
assembler options (by removing the usage of the -S flag that prevented
assembler execution) and also stripping the -Wa, prefix if present
when checking for the test output.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>